(1) Rename and relocate
(2) Fix bug: reference to non-existent `syncManager` object that was factored out in a previous release

(3)  Import initial-storage instead of re-getting storage

(4) Convert updateStatus to arrow function

(5) Insert quota warning after Data header in settings
